+++ rest-client/github_basic.rest
@@ -1,4 +1,4 @@
-POST http://localhost:8080/pubsub/ HTTP/1.1
+POST http://localhost:8080/github/ HTTP/1.1
 content-type: application/json
 
 {
