+++ rest-client/pubsub_basic.rest
@@ -1,0 +1,7 @@
+POST http://localhost:8080/pubsub/ HTTP/1.1
+content-type: application/json
+
+{
+    "message": "sample",
+    "time": "Wed, 21 Oct 2015 18:27:50 GMT"
+}
